packages/backend/requirements.txt
fastapi==0.115.0
slowapi==0.1.9
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3
geoalchemy2==0.15.2
pydantic[email]==2.9.2
pydantic-settings==2.5.2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
celery[redis]==5.4.0
redis==5.1.1
boto3==1.35.36
pillow==10.4.0
httpx==0.27.2
python-multipart==0.0.12
firebase-admin==6.5.0
google-generativeai==0.8.3
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
aiosqlite==0.20.0
greenlet==3.1.1
